pyproject.toml
[tool.poetry]
name = "nonebot-plugin-gsmaterial"
version = "0.1.3"
description = "Genshin daily material plugin for NoneBot2"
authors = ["monsterxcn <monsterxcn@gmail.com>"]
documentation = "https://github.com/monsterxcn/nonebot-plugin-gsmaterial#readme"
license = "MIT"
homepage = "https://github.com/monsterxcn/nonebot-plugin-gsmaterial"
readme = "README.md"
keywords = ["nonebot", "nonebot2", "genshin", "material"]

[tool.poetry.dependencies]
python = ">=3.7.3,<4.0"
nonebot2 = ">=2.0.0a16"
nonebot-adapter-onebot = ">=2.0.0b1"
nonebot-plugin-apscheduler = ">=0.1.0"
httpx = ">=0.20.0, <1.0.0"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"